Moved GeneralErrorScreen and NewTechError into own files; export from index.ts
