+++ src/ensembl/src/shared/error-screen/index.ts
@@ -1,0 +1,2 @@
+export { default as NewTechError } from './NewTechError';
+export { default as GeneralErrorScreen } from './GeneralErrorScreen';
